app/subject/serializers/address.py
from rest_framework import serializers
from subject.models.address import Address


class AddressSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Address
        fields = '__all__'
