requirements.txt
Django
requests
djangorestframework
psycopg2-binary
flake8
drf_spectacular